// src/features/components/factory/hero/types.ts
export enum HeroVariation {
  HERO_ONE = 'HERO_ONE',
  HERO_TWO = 'HERO_TWO',
  HERO_THREE = 'HERO_THREE',
  HERO_FOUR = 'HERO_FOUR',
  HERO_FIVE = 'HERO_FIVE',
  HERO_SIX = 'HERO_SIX',
  HERO_SEVEN = 'HERO_SEVEN',
  HERO_EIGHT = 'HERO_EIGHT',
  HERO_NINE = 'HERO_NINE',
  HERO_TEN = 'HERO_TEN',
  HERO_ELEVEN = 'HERO_ELEVEN',
}

export type IHeroVariation = `${HeroVariation}`;